QoL Update
Quick bugfix that corrects emission not working properly on the chord buttons - I didn't realize that exported projects which use "material.EnableKeyword("_Emission") wouldn't function properly if the Emission keyword of the material was not enabled at the time of export.

Got around this by creating a new material that has emission enabled natively and swapping to it instead of trying to turn the emission on and off via a script.